+++ TidalPy/utilities/lookups/__init__.py
@@ -1,2 +1,2 @@
-from TidalPy.utilities.lookups.intmap3 import IntMap3
-from TidalPy.utilities.lookups.intmap4 import IntMap4
+from TidalPy.utilities.lookups.intmap3 import IntMap3, IntMap3Complex
+from TidalPy.utilities.lookups.intmap4 import IntMap4, IntMap4Complex